requirements.txt
dj-database-url
Django
django-cors-headers
djangorestframework
gunicorn
psycopg2-binary
whitenoise
